# api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/internal/orders/{orderNumber}/status": {
      "put": {
        "operationId": "InternalOrdersController_updateOrderStatus",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update order status by order number (Internal)",
        "tags": [
          "Internal Orders"
        ]
      }
    }
  },
  "info": {
    "title": "Ecommerce-Platform Backend API",
    "description": "",
    "version": "1.0.5",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "pending",
              "processing",
              "shipped",
              "delivered",
              "cancelled",
              "refunded",
              "success"
            ],
            "type": "string",
            "description": "New order status",
            "example": "shipped"
          },
          "trackingNumber": {
            "type": "string",
            "description": "Tracking number for shipped orders",
            "example": "TRACK123456"
          },
          "paymentStatus": {
            "enum": [
              "pending",
              "paid",
              "failed",
              "refunded"
            ],
            "type": "string",
            "description": "Payment status",
            "example": "paid"
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}